refactor: rename database namespace
