+++ Robust.Cdn/Controllers/DownloadCompatibilityController.cs
@@ -1,6 +1,7 @@
-﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
 using Robust.Cdn.Config;
+using Robust.Cdn.DataAccessLayer;
 using Robust.Cdn.Services;
 
 namespace Robust.Cdn.Controllers;
