+++ Robust.Cdn/Jobs/MakeNewManifestVersionsAvailableJob.cs
@@ -1,6 +1,7 @@
-﻿using System.Text.Json;
+using System.Text.Json;
 using Dapper;
 using Quartz;
+using Robust.Cdn.DataAccessLayer;
 using Robust.Cdn.Helpers;
 
 namespace Robust.Cdn.Jobs;
