+++ Robust.Cdn/Services/PublishManager.cs
@@ -1,5 +1,6 @@
-﻿using System.Data.Common;
+using System.Data.Common;
 using Dapper;
+using Robust.Cdn.DataAccessLayer;
 
 namespace Robust.Cdn.Services;
